fixed segfault heredoc expand && created new file heredoc_expand.c
