+++ src/exec/heredoc.c
@@ -6,33 +6,11 @@
 /*   By: csweetin <csweetin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 11:51:17 by cdomet-d          #+#    #+#             */
-/*   Updated: 2024/04/25 18:59:22 by csweetin         ###   ########.fr       */
+/*   Updated: 2024/04/26 15:41:34 by csweetin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "exec.h"
-
-int	heredoc_expand(char **line, t_input *in)
-{
-	char	*temp;
-	int		letter;
-
-	letter = nb_letter(*line, &(in)->env);
-	temp = ft_calloc(sizeof(char), letter + 1);
-	if (!temp)
-		return (1);
-	ft_copy(*line, temp, &(in)->env, 0);
-	free(*line);
-	*line = temp;
-	letter = 0;
-	while (temp[letter])
-	{
-		if (temp[letter] < 0)
-			*line[letter] = temp[letter] * -1;
-		letter++;
-	}
-	return (0);
-}
 
 static void	*h_gnl(int fd, t_input *in)
 {
@@ -54,7 +32,7 @@ static void	*h_gnl(int fd, t_input *in)
 		if (line)
 		{
 			if (in->data[0][0] < 0)
-				if (search_dollar(&line))
+				if (search_dollar_hd(line))
 					if (heredoc_expand(&line, in))
 						return (print_error(errno, NULL));
 			if (write(fd, line, ft_strlen(line)) == -1)
